implementations/django/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-guardian>=2.4.0
django-fsm>=2.8.0
django-simple-history>=3.4.0
django-cors-headers>=4.3.0
python-decouple>=3.8
gunicorn>=21.0.0

# For development - will use SQLite initially
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.3.0